[refactor🔨] : couponIssue를 couponGenerationIssue로 변경함에 따라 관련 파일들 이름변경
