+++ src/main/java/shop/gaship/coupon/coupongenerationissue/adapter/CouponAdapter.java
@@ -1,9 +1,9 @@
-package shop.gaship.coupon.couponissue.adapter;
+package shop.gaship.coupon.coupongenerationissue.adapter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
-import shop.gaship.coupon.couponissue.dto.request.CouponIssueCreationRequestDto;
+import shop.gaship.coupon.coupongenerationissue.dto.request.CouponIssueCreationRequestDto;
 
 /**
  * 쿠폰 기능을 수행하기위해서 외부와의 통신이 필요할때 사용할 adapter class입니다.
@@ -16,7 +16,7 @@ import shop.gaship.coupon.couponissue.dto.request.CouponIssueCreationRequestDto;
 public class CouponAdapter {
     private final WebClient schedulerWebClient;
 
-    public void addCouponIssue(CouponIssueCreationRequestDto couponIssueCreationRequestDto) {
+    public void addCouponGenerationIssue(CouponIssueCreationRequestDto couponIssueCreationRequestDto) {
         schedulerWebClient.post()
             .bodyValue(couponIssueCreationRequestDto)
             .retrieve()
