+++ src/main/java/shop/gaship/coupon/coupongenerationissue/controller/CouponGenerationIssueRestController.java
@@ -1,4 +1,4 @@
-package shop.gaship.coupon.couponissue.controller;
+package shop.gaship.coupon.coupongenerationissue.controller;
 
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.gaship.coupon.couponissue.dto.request.CouponIssueCreationRequestDto;
-import shop.gaship.coupon.couponissue.service.CouponIssueService;
+import shop.gaship.coupon.coupongenerationissue.dto.request.CouponIssueCreationRequestDto;
+import shop.gaship.coupon.coupongenerationissue.service.CouponIssueService;
 
 /**
  * 쿠폰생성발급에 대한 요청을 처리하기위한 controller class입니다.
@@ -19,12 +19,12 @@ import shop.gaship.coupon.couponissue.service.CouponIssueService;
 @RestController
 @RequestMapping("/api/coupon-issues")
 @RequiredArgsConstructor
-public class CouponIssueRestController {
+public class CouponGenerationIssueRestController {
     private final CouponIssueService couponIssueService;
 
     @PostMapping
-    public ResponseEntity<Void> couponIssueAdd(@RequestBody @Valid CouponIssueCreationRequestDto couponIssueCreationRequestDto) {
-        couponIssueService.addCouponIssue(couponIssueCreationRequestDto);
+    public ResponseEntity<Void> couponGenerationIssueAdd(@RequestBody @Valid CouponIssueCreationRequestDto couponIssueCreationRequestDto) {
+        couponIssueService.addCouponGenerationIssue(couponIssueCreationRequestDto);
         return ResponseEntity.ok().build();
     }
 }
