+++ src/main/java/shop/gaship/coupon/coupongenerationissue/repository/CouponIssueRepository.java
@@ -1,7 +1,7 @@
-package shop.gaship.coupon.couponissue.repository;
+package shop.gaship.coupon.coupongenerationissue.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.gaship.coupon.couponissue.entity.CouponGenerationIssue;
+import shop.gaship.coupon.coupongenerationissue.entity.CouponGenerationIssue;
 
 /**
  * @author : 최겸준
